Split linuxTPM -> {trousersTPM,wrappedTPM20} + prefix windows key type with windows
