+++ attest/key_linux.go
@@ -19,19 +19,17 @@ package attest
 import (
 	"fmt"
 
-	"github.com/google/go-tpm/tpm2"
-	"github.com/google/go-tpm/tpmutil"
 	"github.com/google/go-tspi/attestation"
 )
 
-// key12 represents a key bound to a TPM 1.2 device via tcsd.
-type key12 struct {
+// trousersKey12 represents a key bound to a TPM 1.2 device via tcsd.
+type trousersKey12 struct {
 	blob   []byte
 	public []byte
 }
 
-func newKey12(blob, public []byte) ak {
-	return &key12{
+func newTrousersKey12(blob, public []byte) ak {
+	return &trousersKey12{
 		blob:   blob,
 		public: public,
 	}
@@ -39,7 +37,7 @@ func newKey12(blob, public []byte) ak {
 
 // Marshal represents the key in a persistent format which may be
 // loaded at a later time using tpm.LoadKey().
-func (k *key12) marshal() ([]byte, error) {
+func (k *trousersKey12) marshal() ([]byte, error) {
 	out := serializedKey{
 		Encoding:   keyEncodingEncrypted,
 		TPMVersion: TPMVersion12,
@@ -49,12 +47,12 @@ func (k *key12) marshal() ([]byte, error) {
 	return out.Serialize()
 }
 
-func (k *key12) close(tpm tpmBase) error {
+func (k *trousersKey12) close(tpm tpmBase) error {
 	return nil // No state for tpm 1.2.
 }
 
-func (k *key12) activateCredential(tb tpmBase, in EncryptedCredential) ([]byte, error) {
-	t, ok := tb.(*linuxTPM)
+func (k *trousersKey12) activateCredential(tb tpmBase, in EncryptedCredential) ([]byte, error) {
+	t, ok := tb.(*trousersTPM)
 	if !ok {
 		return nil, fmt.Errorf("expected *linuxTPM, got %T", tb)
 	}
@@ -66,8 +64,8 @@ func (k *key12) activateCredential(tb tpmBase, in EncryptedCredential) ([]byte, 
 	return cred, nil
 }
 
-func (k *key12) quote(tb tpmBase, nonce []byte, alg HashAlg) (*Quote, error) {
-	t, ok := tb.(*linuxTPM)
+func (k *trousersKey12) quote(tb tpmBase, nonce []byte, alg HashAlg) (*Quote, error) {
+	t, ok := tb.(*trousersTPM)
 	if !ok {
 		return nil, fmt.Errorf("expected *linuxTPM, got %T", tb)
 	}
@@ -87,104 +85,9 @@ func (k *key12) quote(tb tpmBase, nonce []byte, alg HashAlg) (*Quote, error) {
 	}, nil
 }
 
-func (k *key12) attestationParameters() AttestationParameters {
+func (k *trousersKey12) attestationParameters() AttestationParameters {
 	return AttestationParameters{
 		Public:                  k.public,
 		UseTCSDActivationFormat: true,
-	}
-}
-
-// key20 represents a key bound to a TPM 2.0.
-type key20 struct {
-	hnd tpmutil.Handle
-
-	blob              []byte
-	public            []byte // used by both TPM1.2 and 2.0
-	createData        []byte
-	createAttestation []byte
-	createSignature   []byte
-}
-
-func newKey20(hnd tpmutil.Handle, blob, public, createData, createAttestation, createSig []byte) ak {
-	return &key20{
-		hnd:               hnd,
-		blob:              blob,
-		public:            public,
-		createData:        createData,
-		createAttestation: createAttestation,
-		createSignature:   createSig,
-	}
-}
-
-func (k *key20) marshal() ([]byte, error) {
-	return (&serializedKey{
-		Encoding:   keyEncodingEncrypted,
-		TPMVersion: TPMVersion20,
-
-		Blob:              k.blob,
-		Public:            k.public,
-		CreateData:        k.createData,
-		CreateAttestation: k.createAttestation,
-		CreateSignature:   k.createSignature,
-	}).Serialize()
-}
-
-func (k *key20) close(t tpmBase) error {
-	tpm, ok := t.(*linuxTPM)
-	if !ok {
-		return fmt.Errorf("expected *linuxTPM, got %T", t)
-	}
-	return tpm2.FlushContext(tpm.rwc, k.hnd)
-}
-
-func (k *key20) activateCredential(tb tpmBase, in EncryptedCredential) ([]byte, error) {
-	t, ok := tb.(*linuxTPM)
-	if !ok {
-		return nil, fmt.Errorf("expected *linuxTPM, got %T", tb)
-	}
-
-	ekHnd, _, err := t.getPrimaryKeyHandle(commonEkEquivalentHandle)
-	if err != nil {
-		return nil, err
-	}
-
-	sessHandle, _, err := tpm2.StartAuthSession(
-		t.rwc,
-		tpm2.HandleNull,  /*tpmKey*/
-		tpm2.HandleNull,  /*bindKey*/
-		make([]byte, 16), /*nonceCaller*/
-		nil,              /*secret*/
-		tpm2.SessionPolicy,
-		tpm2.AlgNull,
-		tpm2.AlgSHA256)
-	if err != nil {
-		return nil, fmt.Errorf("creating session: %v", err)
-	}
-	defer tpm2.FlushContext(t.rwc, sessHandle)
-
-	if _, err := tpm2.PolicySecret(t.rwc, tpm2.HandleEndorsement, tpm2.AuthCommand{Session: tpm2.HandlePasswordSession, Attributes: tpm2.AttrContinueSession}, sessHandle, nil, nil, nil, 0); err != nil {
-		return nil, fmt.Errorf("tpm2.PolicySecret() failed: %v", err)
-	}
-
-	return tpm2.ActivateCredentialUsingAuth(t.rwc, []tpm2.AuthCommand{
-		{Session: tpm2.HandlePasswordSession, Attributes: tpm2.AttrContinueSession},
-		{Session: sessHandle, Attributes: tpm2.AttrContinueSession},
-	}, k.hnd, ekHnd, in.Credential[2:], in.Secret[2:])
-}
-
-func (k *key20) quote(tb tpmBase, nonce []byte, alg HashAlg) (*Quote, error) {
-	t, ok := tb.(*linuxTPM)
-	if !ok {
-		return nil, fmt.Errorf("expected *linuxTPM, got %T", tb)
-	}
-	return quote20(t.rwc, k.hnd, tpm2.Algorithm(alg), nonce)
-}
-
-func (k *key20) attestationParameters() AttestationParameters {
-	return AttestationParameters{
-		Public:            k.public,
-		CreateData:        k.createData,
-		CreateAttestation: k.createAttestation,
-		CreateSignature:   k.createSignature,
 	}
 }
